Added a space option and fixed a bug with commented lines.
I changed the "Tab" option to be indicated by a value of 2 spaces. The "Whitespace" option is indicated by an empty string. Thus, the only 2 options that cause the join to be done with tab characters are one which don't have a length of 1. Every other single character is split and joined using that character. This fully preserves previous behavior and makes the labels more accurately descriptive. before this change, the ability to join with space characters was lost.

I realized in my testing the the "startswith" test was being performed per column. It only appeared to work before because I was using column 1 for the replacement test. The side-effect of commented lines' values being joined by a different character is now not an issue either because I moved the startswith test and append up to the conditional that checks the lines to skip.

Files checked in: tools/replaceColumn/replaceColumn.xml